+++ server/controllers/hewanTernakController.js
@@ -1,10 +1,21 @@
 import HewanTernak from "../models/hewanTernakModel.js";
 import Lokasi from "../models/lokasiModel.js";
 import { Sequelize, Op } from "sequelize";
+import Peternak from "../models/peternakModel.js";
 
 export const getHewanTernak = async (req, res) => {
+  const { idPeternak } = req.body;
   try {
-    const response = await HewanTernak.findAll();
+    const response = await HewanTernak.findAll({
+      where: {
+        idPeternak: idPeternak,
+      },
+      include: [
+        {
+          model: Peternak,
+        },
+      ],
+    });
     res.status(200).json(response);
   } catch (error) {
     console.error(error.message);
@@ -46,38 +57,68 @@ export const getHewanTernakByProvinsi = async (req, res) => {
   }
 };
 
-export const getHewanTernakByTahunAndJenis = async (req,res) => {
+export const getHewanTernakByTahunAndJenis = async (req, res) => {
   try {
     const { tahun } = req.body;
     const { jenisHewan } = req.body;
     const result = await HewanTernak.findAll({
       attributes: [
-        [
-          Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"),
-          "Tahun"
-        ],
-        [
-          Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%M')"),
-          "Bulan"
-        ],
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"), "Tahun"],
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%M')"), "Bulan"],
         "jenisHewan",
-        [
-          Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")),
-          "JumlahHewan"
-        ],
+        [Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")), "JumlahHewan"],
       ],
       where: {
         jenisHewan: jenisHewan,
-        [Op.and]: Sequelize.where(Sequelize.fn('YEAR', Sequelize.col('waktuPerubahan')), tahun),
+        [Op.and]: Sequelize.where(
+          Sequelize.fn("YEAR", Sequelize.col("waktuPerubahan")),
+          tahun
+        ),
       },
       group: [
         Sequelize.literal("Tahun"),
         Sequelize.literal("Bulan"),
-        "jenisHewan"
+        "jenisHewan",
       ],
       order: [
         [Sequelize.literal("waktuPerubahan"), "ASC"],
-        ["jenisHewan", "ASC"]
+        ["jenisHewan", "ASC"],
+      ],
+    });
+    res.status(200).json(result);
+  } catch (error) {
+    console.error("Error fetching data:", error);
+    res.status(500).json({ message: "Error fetching data", error });
+  }
+};
+export const getHewanTernakByTahunAndJenisId = async (req, res) => {
+  try {
+    const { tahun } = req.body;
+    const { jenisHewan } = req.body;
+    const { idPeternak } = req.body;
+    const result = await HewanTernak.findAll({
+      attributes: [
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"), "Tahun"],
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%M')"), "Bulan"],
+        "jenisHewan",
+        [Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")), "JumlahHewan"],
+      ],
+      where: {
+        idPeternak: idPeternak,
+        jenisHewan: jenisHewan,
+        [Op.and]: Sequelize.where(
+          Sequelize.fn("YEAR", Sequelize.col("waktuPerubahan")),
+          tahun
+        ),
+      },
+      group: [
+        Sequelize.literal("Tahun"),
+        Sequelize.literal("Bulan"),
+        "jenisHewan",
+      ],
+      order: [
+        [Sequelize.literal("waktuPerubahan"), "ASC"],
+        ["jenisHewan", "ASC"],
       ],
     });
     res.status(200).json(result);
@@ -87,29 +128,49 @@ export const getHewanTernakByTahunAndJenis = async (req,res) => {
   }
 };
 
-export const getHewanTernakByTahun = async (req,res) => {
+export const getHewanTernakByTahun = async (req, res) => {
   try {
     const { tahun } = req.params;
     const result = await HewanTernak.findAll({
       attributes: [
-        [
-          Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"),
-          "Tahun"
-        ],
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"), "Tahun"],
         "jenisHewan",
-        [
-          Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")),
-          "JumlahHewan"
-        ],
+        [Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")), "JumlahHewan"],
       ],
       where: Sequelize.literal(`DATE_FORMAT(waktuPerubahan, '%Y') = ${tahun}`),
-      group: [
-        Sequelize.literal("Tahun"),
-        "jenisHewan"
-      ],
+      group: [Sequelize.literal("Tahun"), "jenisHewan"],
       order: [
         [Sequelize.literal("waktuPerubahan"), "ASC"],
-        ["jenisHewan", "ASC"]
+        ["jenisHewan", "ASC"],
+      ],
+    });
+    res.status(200).json(result);
+  } catch (error) {
+    console.error("Error fetching data:", error);
+    res.status(500).json({ message: "Error fetching data", error });
+  }
+};
+export const getHewanTernakByTahunAndId = async (req, res) => {
+  try {
+    const { tahun } = req.body;
+    const { idPeternak } = req.body;
+    const result = await HewanTernak.findAll({
+      attributes: [
+        [Sequelize.literal("DATE_FORMAT(waktuPerubahan, '%Y')"), "Tahun"],
+        "jenisHewan",
+        [Sequelize.fn("COUNT", Sequelize.col("idHewanTernak")), "JumlahHewan"],
+      ],
+      where: {
+        idPeternak: idPeternak,
+        [Op.and]: Sequelize.where(
+          Sequelize.literal(`DATE_FORMAT(waktuPerubahan, '%Y')`),
+          tahun
+        ),
+      },
+      group: [Sequelize.literal("Tahun"), "jenisHewan"],
+      order: [
+        [Sequelize.literal("waktuPerubahan"), "ASC"],
+        ["jenisHewan", "ASC"],
       ],
     });
     res.status(200).json(result);
@@ -133,9 +194,30 @@ export const getHewanTernakById = async (req, res) => {
 };
 
 export const createHewanTernak = async (req, res) => {
+  const {
+    jenisHewan,
+    famili,
+    berat,
+    jenisKelamin,
+    usia,
+    vaksin,
+    idPeternak,
+    idLokasi,
+    waktuPerubahan,
+  } = req.body;
   try {
-    await HewanTernak.create(req.body);
-    res.status(201).json({ msg: "HewanTernak Created" });
+    const hewanTernak = await HewanTernak.create({
+      jenisHewan: jenisHewan,
+      famili: famili,
+      berat: berat,
+      jenisKelamin: jenisKelamin,
+      usia: usia,
+      vaksin: vaksin,
+      idPeternak: idPeternak,
+      idLokasi: idLokasi,
+      waktuPerubahan: waktuPerubahan,
+    });
+    res.status(201).json(hewanTernak);
   } catch (error) {
     console.error(error.message);
   }
@@ -143,22 +225,66 @@ export const createHewanTernak = async (req, res) => {
 
 export const updateHewanTernak = async (req, res) => {
   try {
-    await HewanTernak.update(req.body, {
+    const hewanTernak = await HewanTernak.findOne({
       where: {
         idHewanTernak: req.params.id,
       },
     });
+
+    if (!hewanTernak) {
+      return res.status(404).json({ msg: "Hewan ternak tidak ditemukan" });
+    }
+
+    const {
+      jenisHewan,
+      famili,
+      berat,
+      jenisKelamin,
+      usia,
+      vaksin,
+      idPeternak,
+      idLokasi,
+      waktuPerubahan,
+    } = req.body;
+    await HewanTernak.update(
+      {
+        jenisHewan: jenisHewan,
+        famili: famili,
+        berat: berat,
+        jenisKelamin: jenisKelamin,
+        usia: usia,
+        vaksin: vaksin,
+        idPeternak: idPeternak,
+        idLokasi: idLokasi,
+        waktuPerubahan: waktuPerubahan,
+      },
+      {
+        where: {
+          idHewanTernak: hewanTernak.idHewanTernak,
+        },
+      }
+    );
     res.status(200).json({ msg: "HewanTernak Updated" });
   } catch (error) {
     console.error(error.message);
+    res.status(500).json({ msg: "Internal Server Error" });
   }
 };
 
 export const deleteHewanTernak = async (req, res) => {
+  const hewanTernak = await HewanTernak.findOne({
+    where: {
+      idHewanTernak: req.params.id,
+    },
+  });
+
+  if (!hewanTernak) {
+    return res.status(404).json({ msg: "Hewan ternak tidak ditemukan" });
+  }
   try {
     await HewanTernak.destroy({
       where: {
-        idHewanTernak: req.params.id,
+        idHewanTernak: hewanTernak.idHewanTernak,
       },
     });
     res.status(200).json({ msg: "HewanTernak Deleted" });
