+++ server/controllers/peternakController.js
@@ -1,3 +1,4 @@
+import bcrypt from "bcrypt";
 import Peternak from "../models/peternakModel.js";
 
 export const getPeternak = async (req, res) => {
@@ -23,9 +24,17 @@ export const getPeternakById = async (req, res) => {
 };
 
 export const createPeternak = async (req, res) => {
+  const {idLokasi, namaLengkap, username, password, nomorTelepon } = req.body;
+  const hashPassword = await bcrypt.hash(password,10);
   try {
-    await Peternak.create(req.body);
-    res.status(201).json({ msg: "Peternak Created" });
+    await Peternak.create({
+      usernamePeternak: username,
+      passwordPeternak: hashPassword,
+      namaPeternak: namaLengkap,
+      nomorTelepon: nomorTelepon,
+      idLokasi: idLokasi,
+    });
+    res.status(201).json({ msg: "Register Berhasil" });
   } catch (error) {
     console.error(error.message);
   }
@@ -33,22 +42,59 @@ export const createPeternak = async (req, res) => {
 
 export const updatePeternak = async (req, res) => {
   try {
-    await Peternak.update(req.body, {
+    const peternak = await Peternak.findOne({
       where: {
         idPeternak: req.params.id,
       },
     });
+
+    if (!peternak) {
+      return res.status(404).json({ msg: 'Peternak tidak ditemukan' });
+    }
+
+    const { namaLengkap, username, password, nomorTelepon } = req.body;
+    let hashPassword;
+
+    if (password === "" || password === null || password === undefined) {
+      hashPassword = peternak.passwordPeternak;
+    } else {
+      console.log("Password received for hashing:",password);
+      hashPassword = await bcrypt.hash(password,10);
+    }
+
+    await Peternak.update(
+      {
+        usernamePeternak: username,
+        passwordPeternak: hashPassword,
+        namaPeternak: namaLengkap,
+        nomorTelepon: nomorTelepon,
+      },
+      {
+        where: {
+          idPeternak: peternak.idPeternak,
+        },
+      }
+    );
     res.status(200).json({ msg: "Peternak Updated" });
   } catch (error) {
-    console.error(error.message);
+    console.error("Error updating peternak:", error.message);
+    res.status(500).json({ msg: "Internal Server Error" });
   }
 };
 
 export const deletePeternak = async (req, res) => {
+  const peternak = await Peternak.findOne({
+    where: {
+      idPeternak: req.params.id,
+    },
+  });
+  if (!peternak) {
+    return res.status(404).json({ msg: "Peternak tidak ditemukan" });
+  }
   try {
     await Peternak.destroy({
       where: {
-        idPeternak: req.params.id,
+        idPeternak: peternak.idPeternak,
       },
     });
     res.status(200).json({ msg: "Peternak Deleted" });
